backend/requirements.txt
Flask
flask-admin
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Migrate
Flask-Mail
Flask-Moment
flask_socketio
Flask-CORS
Flask-Limiter
Flask-Talisman
bcrypt
python-dotenv
numpy
pandas
scipy
scikit-learn
pytesseract
opencv-python
Pillow
APScheduler
plotly
XlsxWriter
Flask-Babel
google-generativeai
email_validator
bleach


